Move test to image
