+++ modal-go/test/image_test.go
@@ -77,3 +77,86 @@ func TestImageFromGcpArtifactRegistry(t *testing.T) {
 	g.Expect(err).ShouldNot(gomega.HaveOccurred())
 	g.Expect(image.ImageId).Should(gomega.HavePrefix("im-"))
 }
+
+func TestCreateOneSandboxTopLevelImageAPI(t *testing.T) {
+	t.Parallel()
+	g := gomega.NewWithT(t)
+	app, err := modal.AppLookup(context.Background(), "libmodal-test", &modal.LookupOptions{CreateIfMissing: true})
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+
+	image := modal.NewImageFromRegistry("alpine:3.21", nil)
+	g.Expect(image.ImageId).Should(gomega.BeEmpty())
+
+	sb, err := app.CreateSandbox(image, nil)
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+	defer sb.Terminate()
+
+	g.Expect(image.ImageId).Should(gomega.HavePrefix("im-"))
+}
+
+func TestCreateOneSandboxTopLevelImageAPISecret(t *testing.T) {
+	t.Parallel()
+	g := gomega.NewWithT(t)
+	app, err := modal.AppLookup(context.Background(), "libmodal-test", &modal.LookupOptions{CreateIfMissing: true})
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+
+	secret, err := modal.SecretFromName(context.Background(), "libmodal-gcp-artifact-registry-test", &modal.SecretFromNameOptions{
+		RequiredKeys: []string{"REGISTRY_USERNAME", "REGISTRY_PASSWORD"},
+	})
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+
+	image := modal.NewImageFromRegistry("us-east1-docker.pkg.dev/modal-prod-367916/private-repo-test/my-image", &modal.ImageFromRegistryOptions{
+		Secret: secret,
+	})
+	g.Expect(image.ImageId).Should(gomega.BeEmpty())
+
+	sb, err := app.CreateSandbox(image, nil)
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+	defer sb.Terminate()
+
+	g.Expect(image.ImageId).Should(gomega.HavePrefix("im-"))
+}
+
+func TestImageFromAwsEcrTopLevel(t *testing.T) {
+	t.Parallel()
+	g := gomega.NewWithT(t)
+
+	app, err := modal.AppLookup(context.Background(), "libmodal-test", &modal.LookupOptions{CreateIfMissing: true})
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+
+	secret, err := modal.SecretFromName(context.Background(), "libmodal-aws-ecr-test", &modal.SecretFromNameOptions{
+		RequiredKeys: []string{"AWS_ACCESS_KEY_ID", "AWS_SECRET_ACCESS_KEY"},
+	})
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+
+	image := modal.NewImageFromAwsEcr("459781239556.dkr.ecr.us-east-1.amazonaws.com/ecr-private-registry-test-7522615:python", secret)
+	g.Expect(image.ImageId).Should(gomega.Equal(""))
+
+	sb, err := app.CreateSandbox(image, nil)
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+	defer sb.Terminate()
+
+	g.Expect(image.ImageId).Should(gomega.HavePrefix("im-"))
+}
+
+func TestImageFromGcpEcrTopLevel(t *testing.T) {
+	t.Parallel()
+	g := gomega.NewWithT(t)
+
+	app, err := modal.AppLookup(context.Background(), "libmodal-test", &modal.LookupOptions{CreateIfMissing: true})
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+
+	secret, err := modal.SecretFromName(context.Background(), "libmodal-gcp-artifact-registry-test", &modal.SecretFromNameOptions{
+		RequiredKeys: []string{"SERVICE_ACCOUNT_JSON"},
+	})
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+
+	image := modal.NewImageFromGcpArtifactRegistry("us-east1-docker.pkg.dev/modal-prod-367916/private-repo-test/my-image", secret)
+	g.Expect(image.ImageId).Should(gomega.Equal(""))
+
+	sb, err := app.CreateSandbox(image, nil)
+	g.Expect(err).ShouldNot(gomega.HaveOccurred())
+	defer sb.Terminate()
+
+	g.Expect(image.ImageId).Should(gomega.HavePrefix("im-"))
+}
